+++ packages/framework/src/Model/Order/OrderInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Shopsys\FrameworkBundle\Model\Order;
+
+interface OrderInterface
+{
+}